refactor: Normalize schema directory
