+++ src/File/Infrastructure/Repositories/FileSqlRepository.ts
@@ -4,7 +4,7 @@ import { ICriteria, IPaginator } from '@digichanges/shared-experience';
 
 import Paginator from '../../../App/Presentation/Shared/Paginator';
 import FileFilter from '../../Presentation/Criterias/FileFilter';
-import FileSchema from '../Schema/FileTypeORM';
+import FileSchema from '../Schemas/FileTypeORM';
 import File from '../../Domain/Entities/File';
 import IFileDomain from '../../InterfaceAdapters/IFileDomain';
