+++ src/Shared/Database/TypeORMCreateConnection.ts
@@ -1,11 +1,10 @@
 import { createConnection } from 'typeorm';
 import { ICreateConnection } from '@digichanges/shared-experience';
-import User from '../../User/Infrastructure/Schema/UserTypeORM';
-import Role from '../../Role/Infrastructure/Schema/RoleTypeORM';
-import Item from '../../Item/Infrastructure/Schema/ItemTypeORM';
-import File from '../../File/Infrastructure/Schema/FileTypeORM';
-import Notification from '../../Notification/Infrastructure/Schema/NotificationTypeORM';
-
+import User from '../../User/Infrastructure/Schemas/UserTypeORM';
+import Role from '../../Role/Infrastructure/Schemas/RoleTypeORM';
+import Item from '../../Item/Infrastructure/Schemas/ItemTypeORM';
+import File from '../../File/Infrastructure/Schemas/FileTypeORM';
+import Notification from '../../Notification/Infrastructure/Schemas/NotificationTypeORM';
 
 class TypeORMCreateConnection implements ICreateConnection
 {
